feat: improved developer guide pages
